pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "civic-extraction"
version = "1.1.0"
description = "CIViC Evidence Extraction using Claude Agent SDK"
requires-python = ">=3.11"

dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "claude-agent-sdk>=1.0.0",
    "PyMuPDF>=1.23.0",  # For PDF page rendering
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
include = [
    "agents*",
    "config*",
    "context*",
    "hooks*",
    "normalization*",
    "prompts*",
    "schemas*",
    "scripts*",
    "tools*",
]
exclude = [
    "data*",
    "logs*",
    "outputs*",
    "tests*",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]